refactor add item actions into seperate file
