+++ src/components/FooterBar.js
@@ -12,22 +12,22 @@ class FooterBar extends Component {
                 style={styles.footerStyle}
                 //onTabChange={(newTabIndex) => alert(`New Tab at position ${newTabIndex}`)}
             >
-            <Tab
-                barBackgroundColor="#37474F"
-                label="Contacts"
-                icon={<Icon size={24} color="white" name="people" />}
-            />
-            <Tab
-                barBackgroundColor="#4f6b99"
-                label="Events"
-                icon={<Icon size={24} color="white" name="view-list" />}
-            />
-            <Tab
-                barBackgroundColor="#00796B"
-                label="Tokens"
-                icon={<Icon size={24} color="white" name="monetization-on" />}
-            />
-          </BottomNavigation>
+                <Tab
+                    barBackgroundColor="#37474F"
+                    label="Contacts"
+                    icon={<Icon size={24} color="white" name="people" />}
+                />
+                <Tab
+                    barBackgroundColor="#4f6b99"
+                    label="Events"
+                    icon={<Icon size={24} color="white" name="view-list" />}
+                />
+                <Tab
+                    barBackgroundColor="#00796B"
+                    label="Tokens"
+                    icon={<Icon size={24} color="white" name="monetization-on" />}
+                />
+            </BottomNavigation>
         );
     }
 }
